+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 - Added GitHub issue template
 - Added more i18n
 - Added auto-play feature
-- Added lodash pacakge
+- Added lodash package
 - Updated audio element in slide show
 - Update toggle hook (useState -> useBoolean hook in Chakra UI)